Re-organize scaled ops translations into common and L2 rules.
Properly separating common ops, independent of the strategy, and L2/Gaussian
scaling ops/translation (which potentially can be altered).
