+++ jax_scaled_arithmetics/lax/__init__.py
@@ -1,3 +1,4 @@
 # Copyright (c) 2023 Graphcore Ltd. All rights reserved.
 from .base_scaling_primitives import set_scaling, set_scaling_p, stop_scaling, stop_scaling_p  # noqa: F401
-from .scaled_ops import *  # noqa: F401, F403
+from .scaled_ops_common import *  # noqa: F401, F403
+from .scaled_ops_l2 import *  # noqa: F401, F403
